% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class.R
\name{a_class.name}
\alias{a_class.name}
\title{Define triple by class, type.}
\arguments{
\item{x}{\code{name}. Likely referring to some variable, i.e., \code{?some_variable}.}

\item{what}{\verb{Character scalar}. Which type to use.}

\item{prefix}{\verb{Character scalar}. Name of the prefix the class is from.
(Default: \code{""})}

\item{...}{additional arguments}
}
\value{
a \code{triple} object.
}
\description{
\code{a_class()} is a helper function to mimic the \verb{a <Class>} pattern
in sparql.
While \code{a_class()} could be called directly, the intended use is to serve as a
template to generate prefix-specific functions for each type known by the
endpoint.
}
\examples{
# a_class defines a triple:
# a_class() handles '?' symbols.
a_class(A, what = "type", prefix = "p")
a_class(?A, what = "type", prefix = "p")
a_class(p_A:A, what = "type", prefix = "p")
}
\seealso{
\link{predicate-methods}
}
