% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{a_type}
\alias{a_type}
\title{Define triple by class type.}
\usage{
a_type(x, ...)
}
\arguments{
\item{x}{object to dispatch on}

\item{...}{additional arguments}
}
\value{
an R object.
}
\description{
\code{a_type()} is a helper function to mimic the \verb{a <Class>} pattern in sparql.
While \code{a_type()} could be called directly, the intended use is to serve as a
template to generate prefix-specific functions for each type known by the
endpoint.
}
\examples{
a_type
}
