% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_type.R
\name{a_type.name}
\alias{a_type.name}
\title{Define triple by class, type.}
\arguments{
\item{x}{\code{name}. Likely referring to some variable, i.e., \code{?some_variable}.}

\item{what}{\verb{Character scalar}. Which type to use.}

\item{...}{additional arguments}
}
\value{
a \code{triple} object.
}
\description{
\code{a_type()} is a helper function to mimic the \verb{a <Class>} pattern
in sparql.
While \code{a_type()} could be called directly, the intended use is to serve as a
template to generate prefix-specific functions for each type known by the
endpoint.
}
\examples{
# a_type defines a triple:
a_type(
    x = var,
    what = c("prefix", "type")
)

# a_type() handles '?' symbols.
a_type(a, what = "")
a_type(?a, what = "")

}
\seealso{
\code{\link[=predicate.formula]{predicate.formula()}}
}
