% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{as.SPARQL}
\alias{as.SPARQL}
\title{Coerce to SPARQL query}
\usage{
as.SPARQL(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{additional argumants.}
}
\value{
a \verb{Character scalar} with a SPARQL query.
}
\description{
Coerce to SPARQL query
}
