% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{OWL_env}
\alias{OWL_env}
\title{S7 class to contain an OWL environment.}
\usage{
OWL_env(x)
}
\arguments{
\item{x}{\code{List} object.}
}
\value{
an \code{OWL_env} object.
}
\description{
\code{OWL_env} is an S7 class which is typically slotted into \code{OWL} objects to
serve as a context, an environment, in which to keep track of the "non-SPARQL
query code" parts of the \code{OWL} object. This includes custom triple generating
functions (\code{term}s), but also the variables named in the query so far.
}
\examples{
OWL_env

}
