% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{predicate}
\alias{predicate}
\title{Define triple by predicate}
\usage{
predicate(x, ...)
}
\arguments{
\item{x}{object to dispatch on}

\item{...}{additional arguments}
}
\value{
an R object.
}
\description{
While \code{predicate()} could be called directly, the intended use
is to serve as a template to generate prefix-specific functions for each
predicate known by the endpoint.
}
\examples{
predicate
}
