% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicate-factory.R
\name{predicate_factory}
\alias{predicate_factory}
\title{Generate custom functions based on available predicates.}
\usage{
predicate_factory(x, to.list = TRUE)
}
\arguments{
\item{x}{\verb{Character scalar}. Names of the desired predicate functions.}

\item{to.list}{\verb{Logical scalar}. Whether to return functions in a list
(\code{TRUE}, Default) or to the environment (\code{FALSE}).}
}
\value{
A list of functions, or to the environment.
}
\description{
Generate custom functions based on available predicates.
}
\examples{
P <- predicate_factory(x = c("predicate_A", "predicate_B", "predicate_C"))

P$predicate_A(subject ~ object)
P$predicate_B(subject ~ object)
P$predicate_C(subject ~ object)

}
