% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{Send a SPARWL query to an endpoint.}
\usage{
send_query(query, path = "sparql/", endpoint_url, out_format = "text/csv")
}
\arguments{
\item{query}{\verb{Character scalar} SPARQL query content.}

\item{endpoint_url}{\verb{Character scalar} sparql endpoint url.}

\item{out_format}{\verb{Character scalar}. (Default: \code{"text/csv"}.)}
}
\description{
Query a SPARQL endpoint with SPARQL.
}
