% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SPARQLquery}
\alias{SPARQLquery}
\title{S7 class to contain a SPARQL query.}
\usage{
SPARQLquery(prefix = list(), query = list(), where = list())
}
\arguments{
\item{prefix}{\verb{Named list of character scalars}, where values are long-form
and names are the corresponding abbreviation.}

\item{query}{\verb{Named list} of character vectors, with first element indicating
query form (SELECT, ASK, DESCRIBE), and the second indicating the content
of the query itself.}

\item{where}{\code{List} of triples}
}
\value{
a \code{SPARQLquery} object.
}
\description{
\code{SPARQLquery} is an S7 class to compose and manage \code{SPARQL} queries.
}
\section{Slots}{

\describe{
\item{\code{prefix}}{\verb{Named list of character scalars}, where values are long-form
and names are the corresponding abbreviation.}

\item{\code{query}}{\verb{Named list} of character vectors, with first element indicating
query form (SELECT, ASK, DESCRIBE), and the second indicating the content
of the query itself.}

\item{\code{where}}{\code{List} of triples}
}}

