% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWL_env-methods.R
\name{tbl_vars.OWL_env}
\alias{tbl_vars.OWL_env}
\title{Column names for RStudio auto-complete}
\arguments{
\item{x}{\code{OWL_env}.}
}
\value{
The names of content within \code{OWL_env} object \code{x}.
}
\description{
Returns the names OWL_env content. Needed for RStudio to complete variable
names
}
