% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{term}
\alias{term}
\title{S7 class to mark domain-specific helper functions.}
\usage{
term(x, hint = "")
}
\arguments{
\item{x}{\code{function} to be decorated to \code{term}.}

\item{hint}{\verb{Character scalar}. Used to print nicer information for generated
functions.}
}
\value{
an object of class \code{term}.
}
\description{
A \code{term} is an S7 class to reflect an RDF term. In RDF, vocabularies are sets
of terms used to describe things. A term is either a class or a property.
In practice, \code{term} is a thin wrapper around \code{function}.
}
\section{Slots}{

\describe{
\item{\code{hint}}{\verb{Character scalar}. Used to print nicer information for generated
functions.}
}}

\examples{
term

}
