% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{triple}
\alias{triple}
\title{S7 class to contain an RDF triple.}
\usage{
triple(subject = c("", ""), predicate = c("", ""), object = c("", ""))
}
\arguments{
\item{subject}{\verb{Character vector}. Length one or two. First position gives
subject name, second (optional, default \code{NULL}) gives prefix.}

\item{predicate}{\verb{Character scalar}. Length one or two. First position gives
predicate name, second (optional, default \code{NULL}) gives prefix.}

\item{object}{\verb{Character scalar}. Length one or two. First position gives
object name, second (optional, default \code{NULL}) gives prefix.}
}
\value{
a \code{triple} object.
}
\description{
\code{triple} is an S7 class contain RDF triples that make up a SPARQL query.
A \code{triple} expression describes a relationship in three parts, known as
\verb{subject–predicate–object} expression. In the data class, this is reflected
by its three slots.
}
\section{Slots}{

\describe{
\item{\code{subject}}{\verb{Character vector}. Length one or two. First position gives
subject name, second (optional, default \code{NULL}) gives prefix.}

\item{\code{predicate}}{\verb{Character scalar}. Length one or two. First position gives
predicate name, second (optional, default \code{NULL}) gives prefix.}

\item{\code{object}}{\verb{Character scalar}. Length one or two. First position gives
object name, second (optional, default \code{NULL}) gives prefix.}
}}

\examples{
triple(
    subject = c("sub_prefix", "subject"),
    predicate = c("pred_prefix", "predicate"),
    object = c("obj_prefix", "object")
)
}
