% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple-methods.R
\name{predicate.triple}
\alias{predicate.triple}
\alias{predicate.pallas::triple}
\title{Define triple by predicate}
\arguments{
\item{x}{\code{triple}.}
}
\value{
\code{predicate()}: The predicate slot of input \code{triple} \code{'x'}.
}
\description{
Define triple by predicate
}
\examples{
x <- triple(
    subject = c("sub_prefix", "subject"),
    predicate = c("pred_prefix", "predicate"),
    object = c("obj_prefix", "object")
)
predicate(x)

}
