% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary-methods.R
\name{vocabulary-methods}
\alias{vocabulary-methods}
\title{Methods for \code{vocabulary} class.}
\arguments{
\item{x}{\code{vocabulary}.}

\item{pattern}{\verb{Character scalar}. Pattern to look for.}
}
\value{
the name of the extra data slot
}
\description{
Methods for \code{vocabulary} class.
}
