% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{where_clause}
\alias{where_clause}
\title{Define one or more triples for a SPARQL WHERE-clause.}
\usage{
where_clause(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{\code{Expression} or sequence of expression separated by comma's, that
each define an RDF triple for the where clause.}
}
\value{
a \code{OWL} with the defined expressions included in the \code{where} slot.
}
\description{
Define one or more triples for a SPARQL WHERE-clause.
}
