% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_endpoint.R, R/where_clause.R
\name{map_endpoint}
\alias{map_endpoint}
\alias{where_clause}
\title{Map out a SPARQL endpoint.}
\usage{
map_endpoint(endpoint_url, strip.blank = TRUE, return.table = FALSE)

where_clause(x, ...)
}
\arguments{
\item{endpoint_url}{\verb{Character scalar} sparql endpoint url.}

\item{strip.blank}{\code{Boolean}. Whether to remove blank nodes from the result.
(Default: \code{TRUE})}

\item{return.table}{\code{Boolean}. Whether to return the raw output table. For
development and debugging. (Default: \code{FALSE}).}

\item{x}{\code{vocabulary} object.}

\item{...}{\code{Expression} or sequence of expression separated by comma's, that
each define an RDF triple for the where clause.}
}
\value{
a data frame with ?domain ?property ?range columns.

a \code{vocabulary} with the defined expressions included in the \code{where}
slot.
}
\description{
Fetch information about known relational information from a sparql endpoint.
}
\examples{
# Don't query endpoints unintentionally.
if(FALSE) {
    x <- map_endpoint("https://sparql.uniprot.org/")
        x |> where_clause()
}

}
