% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_endpoint.R, R/where_clause.R
\name{map_endpoint}
\alias{map_endpoint}
\alias{where_clause.OWL}
\title{Map out a SPARQL endpoint.}
\usage{
map_endpoint(endpoint_url, strip.blank = TRUE, return.table = FALSE)
}
\arguments{
\item{endpoint_url}{\verb{Character scalar} sparql endpoint url.}

\item{strip.blank}{\code{Boolean}. Whether to remove blank nodes from the result.
(Default: \code{TRUE})}

\item{return.table}{\code{Boolean}. Whether to return the raw output table. For
development and debugging. (Default: \code{FALSE}).}
}
\value{
a data frame with ?domain ?property ?range columns.
}
\description{
Fetch information about known relational information from a sparql endpoint.
}
\examples{
# Don't query endpoints unintentionally.
if(FALSE) {

x <- map_endpoint(endpoint_url = "https://sparql.uniprot.org/")
x |>
    where_clause(C.Enzyme(?a)) |>
    where_clause(P.alternativeName(g:h ~ ?j)) |>
    where_clause(P.activity(?d ~ e:f),
                 C.Cluster(b:c))

}

}
