% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_endpoint.R
\name{map_endpoint}
\alias{map_endpoint}
\title{Map out a SPARQL endpoint.}
\usage{
map_endpoint(endpoint_url)
}
\arguments{
\item{endpoint_url}{\verb{Character scalar} sparql endpoint url.}
}
\value{
a data frame with ?domain ?property ?range columns.
}
\description{
Fetch information about known relational information from a sparql endpoint.
}
\examples{
# Don't query endpoints unintentionally.
if(FALSE) {
    map_endpoint("https://sparql.uniprot.org/")
}

}
