% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{OWL}
\alias{OWL}
\title{S7 class to contain a SPARQL query.}
\usage{
OWL(
  .env = list(P = OWL_env(list()), C = OWL_env(list()), V = character(0L), prefixes =
    data.frame()),
  prefix = list(),
  query = list(),
  where = list()
)
}
\arguments{
\item{.env}{\verb{Named list}. Contains convenience functions, for instance
those generated from \code{map_endpoint()}}

\item{prefix}{\verb{Named list of character scalars}, where values are long-form
and names are the corresponding abbreviation.}

\item{query}{\verb{Named list} of character vectors, with first element indicating
query form (SELECT, ASK, DESCRIBE), and the second indicating the content
of the query itself.}

\item{where}{\code{List} of triples}
}
\value{
a \code{OWL} object.
}
\description{
\code{OWL} is an S7 class to compose and manage \code{SPARQL} queries.
}
\section{Slots}{

\describe{
\item{\code{.env}}{\verb{Named list}. Contains convenience functions, for instance
those generated from \code{map_endpoint()}}

\item{\code{prefix}}{\verb{Named list of character scalars}, where values are long-form
and names are the corresponding abbreviation.}

\item{\code{query}}{\verb{Named list} of character vectors, with first element indicating
query form (SELECT, ASK, DESCRIBE), and the second indicating the content
of the query itself.}

\item{\code{where}}{\code{List} of triples}
}}

\examples{
OWL()
}
