% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{OWL}
\alias{OWL}
\title{S7 class to contain a SPARQL query.}
\usage{
OWL(
  .env = list(V = character(0L), prefixes = data.frame()),
  C = list(),
  P = list(),
  prefix = "",
  query = "",
  where = ""
)
}
\arguments{
\item{.env}{\verb{Named list}. Contains additional information, for instance
those generated from \code{map_endpoint()}}

\item{C, P}{\code{OWL_env} objects containing \code{term} functions, for instance
generated by \code{map_endpoint()}.}

\item{prefix}{\verb{Named list of character scalars}, where values are long-form
and names are the corresponding abbreviation.}

\item{query}{\verb{Named list} of character vectors, with first element indicating
query form (SELECT, ASK, DESCRIBE), and the second indicating the content
of the query itself.}

\item{where}{\code{List} of triples}
}
\value{
a \code{OWL} object.
}
\description{
\code{OWL} is an S7 class to compose and manage \code{SPARQL} queries. \code{OWL} objects.
Have an attached environment, in which to keep track of the "non-SPARQL
query code" parts of the \code{OWL} object. This includes custom triple generating
functions (\code{term}s), but also the variables named in the query so far.
}
\section{Slots}{

\describe{
\item{\code{.env}}{\verb{Named list}. Contains additional envitonmental information
besides convenience functions generated from \code{map_endpoint()}}

\item{\code{sparql.}}{\verb{Named list}. Contains SPARQL code in three sections:
\code{prefix}, \code{query} and \code{where}.}
}}

\examples{
OWL()
}
