% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWL-methods.R
\name{OWL-methods}
\alias{OWL-methods}
\title{Methods for \code{OWL} class.}
\arguments{
\item{x}{\code{OWL}.}

\item{pattern}{\verb{Character scalar}. Pattern to look for.}
}
\value{
the name of the extra data slot
}
\description{
Methods for \code{OWL} class.
}
