% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicate.R
\name{predicate.formula}
\alias{predicate.formula}
\title{Define triple by predicate}
\arguments{
\item{x}{\code{formula}. Describing the relationship \code{subject ~ object}.}

\item{what}{\verb{Character scalar}. Which predicate to use.}

\item{...}{additional arguments}
}
\value{
a \code{triple} object.
}
\description{
While \code{predicate()} could be called directly, the intended use
is to serve as a template to generate prefix-specific functions for each
predicate known by the endpoint.
}
\examples{
# predicate defines a triple:
predicate(
    x = sub_prefix:subject ~ ob_prefix:object,
    what = c("prefix", "predicate")
)

# predicate() handles '?' symbols (Unlike regular formula objects).
predicate(a:b~c:d, what = "")
predicate( ?b~c:d, what = "")
predicate(a:b~ ?d, what = "")
predicate( ?b~ ?d, what = "")

}
