% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{select_query}
\alias{select_query}
\alias{select_clause}
\title{Define a SELECT SPARQL command.}
\usage{
select_clause(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{\code{Expression} the SELECT statement.}
}
\value{
a \code{vocabulary} with the defined expressions included in the \code{query}
slot.
}
\description{
Define a SELECT SPARQL command.
}
